requirements.txt
dj-database-url==0.4.2
Django==2.0
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.3.2
pytz==2017.3
whitenoise==3.2
